Pull out the full setup crap into a util
